pyproject.toml
[tool.poetry]
name = "ualr-chatbot"
version = "0.1.0"
description = ""
authors = ["Ishmam <ishmam.a.solaiman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
requests = "^2.32.3"
openpyxl = "^3.1.5"
sentence-transformers = "^4.1.0"
faiss-cpu = "^1.10.0"
streamlit = "^1.44.1"
beautifulsoup4 = "^4.13.4"
selenium = "^4.31.0"
webdriver-manager = "^4.0.2"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
